# Conflicts:
#	src/main/java/iuh/fit/se/techgalaxy/exception/ErrorCode.java

+++ src/main/java/iuh/fit/se/techgalaxy/service/impl/ProductVariantDetailServiceImpl.java
@@ -108,6 +108,10 @@ public class ProductVariantDetailServiceImpl implements ProductVariantDetailServ
     @Override
     public void deleteProductVariantDetail(String productDetailId) {
         try {
+            long usageCount = productVariantDetailRepository.countOrderDetailsByProductVariantDetailId(productDetailId);
+            if (usageCount > 0) {
+                throw new AppException(ErrorCode.PRODUCT_DELETE_FAILED);
+            }
             productVariantDetailRepository.deleteById(productDetailId);
         } catch (Exception e) {
             throw new AppException(ErrorCode.PRODUCT_DELETE_FAILED);
@@ -131,5 +135,21 @@ public class ProductVariantDetailServiceImpl implements ProductVariantDetailServ
     public ProductDetailResponse findProductVariantDetailByProductVariantAndColorAndMemory(String productVariantId, String color, String memory) {
         ProductVariantDetail productVariantDetail = productVariantDetailRepository.findProductVariantDetailByProductVariantAndColorAndMemory(productVariantId, color, memory);
         return productVariantDetailMapper.toResponse(productVariantDetail);
+    }
+
+    @Override
+    public void updateQuantity(String productVariantDetailId, int quantity) {
+        ProductVariantDetail productVariantDetail = productVariantDetailRepository.findById(productVariantDetailId).orElseThrow(() ->
+                new AppException(ErrorCode.PRODUCT_NOTFOUND));
+        if (productVariantDetail.getQuantity() >= quantity) {
+            productVariantDetail.setQuantity(productVariantDetail.getQuantity() - quantity);
+        } else {
+            throw new AppException(ErrorCode.PRODUCT_NOT_ENOUGH);
+        }
+        try {
+            productVariantDetailRepository.save(productVariantDetail);
+        } catch (Exception e) {
+            throw new AppException(ErrorCode.PRODUCT_UPDATE_FAILED);
+        }
     }
 }
